requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

PyPDF2==3.0.1
pdfplumber==0.10.3

easyocr==1.7.0

playwright==1.40.0

sentence-transformers==5.2.0
transformers==4.57.3
huggingface-hub==0.36.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
numpy==1.24.3
pillow==10.1.0

gunicorn==21.2.0
requests==2.31.0
httpx==0.25.2
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1

      